Add files via upload
# Stock Analysis VBA Project
## Overview of Project
During this project, VBA and its functions were used to analyze stocks for Steve. Steve is a recent college graduate who majored in Finance, his parents (who prefer to invest in alternative energy companies' stocks), were so proud of him that they opted to become his first clients. Steve sought out professional help to assist him with analyzing the data he gathered about various publicly traded companies which specialize in alternative energy.
## Results

  
The illustration presented  is a table depicting the returns of the listed companies (indicated by their tickers) for the year 2017. This data shows that it was overall a good year for the featured alternative energy companies, particularily DQ. This data also shows us that some companies made reletively little profit, while TERP stocks resulted in a loss on returns. Based on the analysis of 2017 alone, one might be interested in investing in DQ.
A good analyst would be remiss to forgo analyzing broader data, especially when it comes to conducting a financial analysis. The illistration presented below will show that allbut two of the alternative energy companies listed returned gains in 2018. The biggest winner being ENPH, and the biggest losser (in sharp contrast to 2017) being DQ.
 
## Summary
My (14-year-old son’s) laptop wouldn’t allow me to download the Challenge starter code; therefore, I was unable to refactor it. The photo listed below was the error message which came up when I attempted to download the code. After researching the matter online, it seems that there may be malware on this device, and the workarounds a above my comprehension.
 
After looking through the challenge instructions, I tried to create a second module from the VBA code I had originally been working on for the class. This proved futile as all changes made to Module 2 were reflected on Module 1. I deleted Module two. However, as I looked through the project’s instructions, I realized that the data from the work I had been doing was all but the same as the data depicted as a checkpoint. This made me decide to at the very least submit my findings and prove that I am capable of doing this work.
I believe that I have an understanding of what refactoring would entail as I did attempt to do so on a different Module. I suppose the advantage of refactoring the code is that it would be much more seamless than the incremental macros I wrote as I was following the lesson. I can’t think of any disadvantage other than possibly of making the mistake of not editing the cells/range and worksheets with the respective data.
The original code (in my case the only code), has the disadvantage of being incremental. I wrote new code as I progressed in this lesson, and although I tried to be as detailed as possible, it is lengthy code to cipher through. Although I would say that its advantage is that it may be easier to pinpoint where an error was made and debug it.